requirements/requirements.txt
appdirs==1.4.4
arcgis==1.8.5
asn1crypto @ file:///tmp/build/80754af9/asn1crypto_1596577642040/work
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
azure-core @ file:///C:/bin/miniconda_4.3.21_x64/conda-bld/azure-core_1615482738248/work
azure-storage-blob @ file:///C:/bin/miniconda_4.3.21_x64/conda-bld/azure-storage-blob_1615483038755/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
black==19.10b0
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
blinker==1.4
brotlipy==0.7.0
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools==4.2.2
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247308275/work
cftime==1.0.0b1
chardet @ file:///C:/ci/chardet_1607706910910/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography==2.8
cycler==0.10.0
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
despatch==0.1.0
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
flake8 @ file:///tmp/build/80754af9/flake8_1615834841867/work
future==0.18.2
google-auth==1.33.1
google-auth-oauthlib==0.4.4
gspread==3.7.0
h5py==2.10.0
html5lib @ file:///tmp/build/80754af9/html5lib_1593446221756/work
httplib2==0.19.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1615900693889/work
iniconfig @ file:///home/linux1/recipes/ci/iniconfig_1610983019677/work
ipykernel @ file:///C:/Users/bri11091/conda-bld/ipykernel_1611123597796/work/dist/ipykernel-5.1.1-py3-none-any.whl
ipython @ file:///C:/bin/miniconda_4.3.21_x64/conda-bld/ipython_1616702731266/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.4.2
isodate @ file:///C:/bin/miniconda_4.3.21_x64/conda-bld/isodate_1615438109370/work
jdcal==1.4.1
jedi @ file:///C:/bin/miniconda_4.3.21_x64/conda-bld/jedi_1616689496604/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
json5==0.9.4
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///C:/Users/bri11091/conda-bld/jupyter_client_1599780818163/work
jupyter-console @ file:///C:/apps/c/conda-bld/jupyter_console_1603313573079/work
jupyter-contrib-core @ file:///C:/Users/bri11091/conda-bld/jupyter_contrib_core_1595953007349/work
jupyter-contrib-nbextensions @ file:///C:/apps/conda/conda-bld/jupyter_contrib_nbextensions_1600372176008/work
jupyter-core @ file:///C:/Users/bri11091/conda-bld/jupyter_core_1595863163509/work
jupyter-highlight-selected-word @ file:///C:/Users/bri11091/conda-bld/jupyter_highlight_selected_word_1595962986040/work
jupyter-latex-envs @ file:///C:/Users/bri11091/conda-bld/jupyter_latex_envs_1595955083713/work
jupyter-nbextensions-configurator @ file:///C:/Users/bri11091/conda-bld/jupyter_nbextensions_configurator_1595955508015/work
jupyterlab==2.2.7
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
keyring @ file:///C:/Users/bri11091/conda-bld/keyring_1599850491448/work
kiwisolver @ file:///C:/ci/kiwisolver_1612282618948/work
lerc @ file:///Volumes/JetDrive/virtualenv/miniconda/conda-bld/lerc_1594078116229/work
lxml @ file:///C:/ci/lxml_1616443391272/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///C:/apps/conda/conda-bld/matplotlib_1599707533873/work
mccabe==0.6.1
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.3.0
mkl-random==1.2.0
mkl-service==2.3.0
mpmath==1.2.1
msrest @ file:///C:/bin/miniconda_4.3.21_x64/conda-bld/msrest_1615480661627/work
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat @ file:///C:/Users/bri11091/conda-bld/nbformat_1600098806325/work
netCDF4==1.5.4
networkx @ file:///C:/apps/conda/conda-bld/networkx_1605153751913/work
nose @ file:///tmp/build/80754af9/nose_1606773131901/work
notebook @ file:///C:/ci/notebook_1595933619049/work
ntlm-auth==1.4.0
numexpr @ file:///C:/ci/numexpr_1614798720432/work
numpy @ file:///C:/bin/miniconda_4.3.21_x64/conda-bld/numpy_and_numpy_base_1616785993847/work
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
openpyxl @ file:///tmp/build/80754af9/openpyxl_1615411699337/work
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas @ file:///C:/ci/pandas_1614711338358/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pathspec==0.7.0
pefile==2019.4.18
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ProgramData/Miniconda3/conda-bld/pillow-simd_1605046977129/work
pluggy @ file:///C:/ci/pluggy_1615976530170/work
pro-notebook-integration @ file:///C:/Users/bri11091/conda-bld/pro_notebook_integration_1606951686363/work
prometheus-client @ file:///C:/Users/bri11091/conda-bld/prometheus_client_1600199393947/work
prompt-toolkit==3.0.5
psutil @ file:///C:/ci/psutil_1612298033174/work
py @ file:///tmp/build/80754af9/py_1607971587848/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle @ file:///home/ktietz/src/ci_mi/pycodestyle_1612807597675/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes @ file:///home/ktietz/src/ci_ipy2/pyflakes_1612551159640/work
Pygments @ file:///C:/Users/bri11091/conda-bld/pygments_1600123308736/work
PyJWT @ file:///C:/ci/pyjwt_1610911411733/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
pyshp @ file:///tmp/build/80754af9/pyshp_1610641829514/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
pytest==0.0.0
python-certifi-win32 @ file:///C:/Users/davi9349/Anaconda3/conda-bld/python-certifi-win32_1601333401516/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz==2020.1
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq @ file:///C:/Users/bri11091/conda-bld/pyzmq_1600288036950/work
regex @ file:///C:/ci/regex_1616096096031/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-kerberos==0.12.0
requests-negotiate-sspi @ file:///C:/Users/davi9349/Anaconda3/conda-bld/requests-negotiate-sspi_1593212662136/work
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
saspy @ file:///C:/bin/miniconda_4.3.21_x64/conda-bld/saspy_1612928306700/work
scipy @ file:///C:/ci/scipy_1616703425152/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
simplegeneric==0.8.1
six==1.15.0
swat @ file:///C:/ProgramData/Anaconda3/conda-bld/swat_1611276133126/work
sympy==1.5.1
terminado==0.9.3
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets==4.3.3
typed-ast @ file:///C:/ci/typed-ast_1610484654578/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
ujson @ file:///C:/ci/ujson_1611244941645/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
wincertstore==0.2
winkerberos==0.7.0
wrapt==1.12.1
x86cpu==0.4
xarray @ file:///tmp/build/80754af9/xarray_1614625053385/work
xlrd==1.2.0
xlwt==1.3.0
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
